+++ frontend/components/modals/UnstakedModal.tsx
@@ -1,31 +1,31 @@
 import { PythBalance } from '@pythnetwork/staking'
 import { BaseModal } from './BaseModal'
 
-export type UnlockedModalProps = {
-  isUnlockedModalOpen: boolean
-  setIsUnlockedModalOpen: (open: boolean) => void
+export type UnstakedModalProps = {
+  isUnstakedModalOpen: boolean
+  setIsUnstakedModalOpen: (open: boolean) => void
   unlockedPythBalance?: PythBalance
   unlockingPythBalance?: PythBalance
 }
-export function UnlockedModal({
-  isUnlockedModalOpen,
-  setIsUnlockedModalOpen,
+export function UnstakedModal({
+  isUnstakedModalOpen,
+  setIsUnstakedModalOpen,
   unlockedPythBalance,
   unlockingPythBalance,
-}: UnlockedModalProps) {
+}: UnstakedModalProps) {
   return (
     <BaseModal
-      isModalOpen={isUnlockedModalOpen}
-      setIsModalOpen={setIsUnlockedModalOpen}
-      title={'Unlocked Tokens'}
+      isModalOpen={isUnstakedModalOpen}
+      setIsModalOpen={setIsUnstakedModalOpen}
+      title={'Unstaked Tokens'}
     >
       <p className="mb-6 leading-6">
-        Unlocking tokens enables you to withdraw them from the program after a
-        cooldown period of two epochs of which they become unlocked tokens.
-        Unlocked tokens cannot participate in governance.
+        Unstaking tokens enables you to withdraw them from the program after a
+        cooldown period of two epochs of which they become unstaked tokens.
+        Unstaked tokens cannot participate in governance.
       </p>
       <p className="leading-6">
-        You currently have {unlockedPythBalance?.toString()} unlocked tokens.
+        You currently have {unlockedPythBalance?.toString()} unstaked tokens.
       </p>
       {unlockingPythBalance && !unlockingPythBalance.isZero() ? (
         <p className="mt-4 leading-6">
